Fix: Minor fixes and performance improvements

+++ app/src/main/java/app/africa/autocheck/core/data/cars/CarDetails.kt
@@ -18,14 +18,14 @@ data class CarDetails(
     val gradeScore: Double?,
 
     val hasWarranty: Boolean,
-    val marketplacePrice: Int,
-    val marketplaceOldPrice: Int,
+    val marketplacePrice: Double,
+    val marketplaceOldPrice: Double,
     val hasFinancing: Boolean,
     val mileageUnit: String?,
 
-    val installment: Int,
+    val installment: Double,
     val depositReceived: Boolean,
-    val loanValue: Int,
+    val loanValue: Double,
     val websiteUrl: String?,
     val bodyTypeId: Int,
     val sold: Boolean,
